refactor(api): modularize providers/[id]/models route into handler-per-provider (#3594)
Break the 2589-line models route into 36 focused modules:
- models/config.ts — provider configuration
- models/handleGetModels.ts — main GET handler
- models/handlers/* — per-provider model handlers (20 files)
- models/customNormalizers.ts — normalization utilities
- models/antigravityHelpers.ts — antigravity-specific helpers

Pure extraction, no behavioral changes. Supersedes #4419. Part of #3594.
