+++ src/app/api/providers/[id]/models/handlers/index.ts
@@ -1,0 +1,23 @@
+export { handleNoAuth } from "./noauth";
+export { handleBedrock } from "./bedrock";
+export { handleOpenAiCompatible, handleAnthropicCompatible } from "./openai-compat";
+export {
+  handleAzureAi,
+  handleAzureOpenAi,
+  handleWatsonx,
+  handleOci,
+  handleSap,
+  handleDataRobot,
+} from "./enterprise";
+export { handleVertex, handleGeminiCli, handleGlm } from "./google";
+export { handleGithub, handleKiro } from "./copilot";
+export {
+  handleAntigravity,
+  handleCursor,
+  handleInnerAi,
+  handleClaude,
+  handleReka,
+  handleQwenOauth,
+} from "./special";
+export { handleGenericConfig } from "./generic";
+export type { HandlerContext, SyncedModel, FallbackOpts, ResponsePayload } from "./types";
